Item count parameter can be up to long value.

+++ L2J_Mobius_04.0_GrandCrusade/java/org/l2jmobius/gameserver/data/ItemTable.java
@@ -378,7 +378,7 @@ public class ItemTable
 		return item;
 	}
 	
-	public Item createItem(String process, int itemId, int count, Player actor)
+	public Item createItem(String process, int itemId, long count, Player actor)
 	{
 		return createItem(process, itemId, count, actor, null);
 	}
