+++ L2J_Mobius_Classic_2.7_Antharas/java/org/l2jmobius/gameserver/data/ItemTable.java
@@ -379,7 +379,7 @@ public class ItemTable
 		return item;
 	}
 	
-	public Item createItem(String process, int itemId, int count, Player actor)
+	public Item createItem(String process, int itemId, long count, Player actor)
 	{
 		return createItem(process, itemId, count, actor, null);
 	}
